+++ CHANGELOG.md
@@ -40,7 +40,7 @@
 * If you need to customize the appearance of the user location indicator, you can subclass `UserPuckCourseView` and `UserHaloCourseView` as a starting point. ([#2968](https://github.com/mapbox/mapbox-navigation-ios/pull/2968))
 * Fixed an issue where route line disappears when changing `MapView` style. ([#3136](https://github.com/mapbox/mapbox-navigation-ios/pull/3136))
 * Added `NavigationOptions.navigationMapView` property to allow customization or reusing possibilities for `NavigationViewController.navigationMapView` ([#3186](https://github.com/mapbox/mapbox-navigation-ios/pull/3186)).
-* Improved the performance for vanishing route line by reducing calculation for route line update. Replaced `NavigationMapView.updateRoute(_:)` with `NavigationMapView.updateVanishingRouteLine(coordinate:)` to update the vanishing route line, and keep the route line vanishing point synced with the user location. ([#3201](https://github.com/mapbox/mapbox-navigation-ios/pull/3201)).
+* Renamed the `NavigationMapView.updateRoute(_:)` method to `NavigationMapView.travelAlongRouteLine(to:)`. Improved the performance of updating the route line to change color at the user’s location as they progress along the route. ([#3201](https://github.com/mapbox/mapbox-navigation-ios/pull/3201)).
 
 ### Location tracking